+++ ios/Classes/YandexMapController.swift
@@ -15,9 +15,6 @@ public class YandexMapController: NSObject, FlutterPlatformView {
   private var polylines: [YMKPolylineMapObject] = []
   private var polygons: [YMKPolygonMapObject] = []
   public let mapView: YMKMapView
-  
-  private var searchManager: YMKSearchManager?
-  private var suggestSession: YMKSearchSuggestSession?
 
   public required init(id: Int64, frame: CGRect, registrar: FlutterPluginRegistrar) {
     self.pluginRegistrar = registrar
@@ -30,8 +27,6 @@ public class YandexMapController: NSObject, FlutterPlatformView {
     self.userLocationLayer =
                 YMKMapKit.sharedInstance().createUserLocationLayer(with: mapView.mapWindow)
     super.init()
-    
-    searchManager = YMKSearch.sharedInstance().createSearchManager(with: .combined)
     
     self.mapView.mapWindow.map.addCameraListener(with: self)
     
@@ -84,19 +79,16 @@ public class YandexMapController: NSObject, FlutterPlatformView {
       removePolygon(call)
       result(nil)
     case "zoomIn":
-        zoomIn()
-        result(nil)
+      zoomIn()
+      result(nil)
     case "zoomOut":
-        zoomOut()
-        result(nil)
+      zoomOut()
+      result(nil)
     case "getTargetPoint":
-        let targetPoint = getTargetPoint()
-        result(targetPoint)
+      let targetPoint = getTargetPoint()
+      result(targetPoint)
     case "moveToUser":
       moveToUser()
-      result(nil)
-    case "getSuggestions":
-      getSuggestions(call)
       result(nil)
     default:
       result(FlutterMethodNotImplemented)
@@ -131,34 +123,34 @@ public class YandexMapController: NSObject, FlutterPlatformView {
     map.setMapStyleWithStyle(params["style"] as! String)
   }
     
-    public func zoomIn() {
-        zoom(1)
-    }
-    
-    public func zoomOut() {
-        zoom(-1)
-    }
-    
-    private func zoom(_ step: Float) {
-        let point = mapView.mapWindow.map.cameraPosition.target
-        let zoom = mapView.mapWindow.map.cameraPosition.zoom
-        let azimuth = mapView.mapWindow.map.cameraPosition.azimuth
-        let tilt = mapView.mapWindow.map.cameraPosition.tilt
-        let currentPosition = YMKCameraPosition(
-            target: point,
-            zoom: zoom+step,
-            azimuth: azimuth,
-            tilt: tilt
-         )
-        mapView.mapWindow.map.move(
-            with: currentPosition,
-            animationType: YMKAnimation(
-                type: YMKAnimationType.smooth,
-                duration: 1
-                ),
-            cameraCallback: nil
+  public func zoomIn() {
+      zoom(1)
+  }
+  
+  public func zoomOut() {
+      zoom(-1)
+  }
+  
+  private func zoom(_ step: Float) {
+      let point = mapView.mapWindow.map.cameraPosition.target
+      let zoom = mapView.mapWindow.map.cameraPosition.zoom
+      let azimuth = mapView.mapWindow.map.cameraPosition.azimuth
+      let tilt = mapView.mapWindow.map.cameraPosition.tilt
+      let currentPosition = YMKCameraPosition(
+          target: point,
+          zoom: zoom+step,
+          azimuth: azimuth,
+          tilt: tilt
         )
-    }
+      mapView.mapWindow.map.move(
+          with: currentPosition,
+          animationType: YMKAnimation(
+              type: YMKAnimationType.smooth,
+              duration: 1
+              ),
+          cameraCallback: nil
+      )
+  }
 
   public func move(_ call: FlutterMethodCall) {
     let params = call.arguments as! [String: Any]
@@ -333,70 +325,6 @@ public class YandexMapController: NSObject, FlutterPlatformView {
       mapObjects.remove(with: polygon)
       polygons.remove(at: polygons.firstIndex(of: polygon)!)
     }
-  }
-
-  public func getSuggestions(_ call: FlutterMethodCall) {
-    let params = call.arguments as! [String: Any]
-    let boundingBox = YMKBoundingBox.init(southWest: YMKPoint.init(latitude: params["southWestLatitude"] as! Double,
-                                                                   longitude: params["southWestLongitude"] as! Double),
-                                          northEast: YMKPoint.init(latitude: params["northEastLatitude"] as! Double,
-                                                                   longitude: params["northEastLongitude"] as! Double))
-    let formattedAddress = params["formattedAddress"] as! String
-    
-    let responseHandler = {(searchResponse: [YMKSuggestItem]?, error: Error?) -> Void in
-      if searchResponse != nil {
-        var suggestItems = [Any]()
-        for suggestItem in searchResponse! {
-          var dict = [String : Any]()
-          dict["title"] = suggestItem.title.text
-          if (suggestItem.subtitle != nil) {
-            dict["subtitle"] = suggestItem.subtitle?.text
-          }
-          if (suggestItem.displayText != nil) {
-            dict["displayText"] = suggestItem.displayText
-          }
-          dict["searchText"] = suggestItem.searchText;
-          dict["tags"] = suggestItem.tags;
-          var suggestItemType = String()
-          switch suggestItem.type {
-          case .toponym:
-              suggestItemType = "TOPONYM"
-          case .business:
-              suggestItemType = "BUSINESS"
-          case .transit:
-              suggestItemType = "TRANSIT"
-          default:
-              suggestItemType = "UNKNOWN"
-          }
-          dict["type"] = suggestItemType;
-          suggestItems.append(dict)
-        }
-        let arguments: [String:Any?] = [
-          "suggests": suggestItems
-        ]
-        self.methodChannel.invokeMethod("onSuggestResponseTitles", arguments: arguments)
-      }
-    }
-
-    if suggestSession == nil {
-      suggestSession = searchManager?.createSuggestSession()
-    }
-    
-    var suggestType = YMKSuggestType()
-    switch params["suggestType"] as! String {
-    case "GEO":
-      suggestType = YMKSuggestType.geo
-    case "BIZ":
-      suggestType = YMKSuggestType.biz
-    case "TRANSIT":
-      suggestType = YMKSuggestType.transit
-    default:
-      suggestType = YMKSuggestType.init(rawValue: 0)
-    }
-    let suggestOptions = YMKSuggestOptions.init(suggestTypes: suggestType,
-                                                userPosition: nil,
-                                                suggestWords: params["suggestWords"] as! Bool)
-    suggestSession?.suggest(withText: formattedAddress, window: boundingBox, suggestOptions: suggestOptions, responseHandler: responseHandler)
   }
 
   public func moveToUser() {
